posts overlay section broken down into components
