+++ packages/desktop/src/components/post/components/feed/index.js
@@ -1,0 +1,3 @@
+import Feed from './feed';
+
+export default Feed;
