+++ packages/desktop/src/components/post/components/newTopic/index.js
@@ -1,0 +1,3 @@
+import NewTopic from './newTopic';
+
+export default NewTopic;
